refactor: rename vue-directives
