+++ packages/web/element-plus/index.ts
@@ -1,7 +1,7 @@
 import installer from './defaults'
 export * from '@element-plus/components'
 export * from '@element-plus/constants'
-export * from '@element-plus/directives'
+export * from '@element-plus/vue-directives'
 export * from '@element-plus/hooks'
 export * from '@element-plus/tokens'
 export * from './make-installer'
