+++ packages/web/vue-directives/src/index.ts
